refactor: rename experiments to Editorial Notes and Editorial Updates
Review Notes → Editorial Notes, Refine Notes → Editorial Updates.
Full rename: PHP namespaces, class names, feature slugs, DB option keys,
Webpack entries, TS source dirs, REST ability routes, plugin IDs,
integration tests, E2E specs, mock fixtures, and docs.

V1_0_0 upgrade migration copies existing enabled-state options from
old keys to new keys on plugin upgrade.
