+++ src/experiments/editorial-updates/index.tsx
@@ -1,0 +1,24 @@
+/**
+ * Editorial Updates Experiment.
+ */
+
+/**
+ * WordPress dependencies
+ */
+import { PostTypeSupportCheck } from '@wordpress/editor';
+import { registerPlugin } from '@wordpress/plugins';
+
+/**
+ * Internal dependencies
+ */
+import EditorialUpdatesPlugin from './components/EditorialUpdatesPlugin';
+
+if ( ( window as any ).aiEditorialUpdatesData?.enabled ) {
+	registerPlugin( 'ai-editorial-updates', {
+		render: () => (
+			<PostTypeSupportCheck supportKeys="editor.notes">
+				<EditorialUpdatesPlugin />
+			</PostTypeSupportCheck>
+		),
+	} );
+}
