+++ tests/e2e-request-mocking/responses/OpenAI/editorial-notes-completions.json
