+++ tests/e2e-request-mocking/responses/OpenAI/editorial-notes-responses.json
