+++ tests/e2e-request-mocking/responses/OpenAI/editorial-updates-completions.json
@@ -1,0 +1,36 @@
+{
+	"id": "chatcmpl-EditorialUpdatesTest",
+	"object": "chat.completion",
+	"created": 1772133006,
+	"model": "gpt-4o-mini-2024-07-18",
+	"choices": [
+		{
+			"index": 0,
+			"message": {
+				"role": "assistant",
+				"content": "This is the refined block content.",
+				"refusal": null,
+				"annotations": []
+			},
+			"logprobs": null,
+			"finish_reason": "stop"
+		}
+	],
+	"usage": {
+		"prompt_tokens": 100,
+		"completion_tokens": 20,
+		"total_tokens": 120,
+		"prompt_tokens_details": {
+			"cached_tokens": 0,
+			"audio_tokens": 0
+		},
+		"completion_tokens_details": {
+			"reasoning_tokens": 0,
+			"audio_tokens": 0,
+			"accepted_prediction_tokens": 0,
+			"rejected_prediction_tokens": 0
+		}
+	},
+	"service_tier": "default",
+	"system_fingerprint": "fp_mocked"
+}
