--- tests/e2e-request-mocking/responses/OpenAI/refine-notes-responses.json
@@ -1,68 +1,0 @@
-{
-	"id": "resp_MockRefineNotes",
-	"object": "response",
-	"created_at": 1772806547,
-	"status": "completed",
-	"background": false,
-	"billing": {
-		"payer": "developer"
-	},
-	"completed_at": 1772806548,
-	"error": null,
-	"frequency_penalty": 0.0,
-	"incomplete_details": null,
-	"instructions": "You are an editorial assistant for WordPress. Your task is to update a single block...",
-	"max_output_tokens": null,
-	"max_tool_calls": null,
-	"model": "gpt-4o-mini-2024-07-18",
-	"output": [
-		{
-			"id": "msg_MockRefineNotesMsg",
-			"type": "message",
-			"status": "completed",
-			"content": [
-				{
-					"type": "output_text",
-					"annotations": [],
-					"logprobs": [],
-					"text": "This is the refined block content."
-				}
-			],
-			"role": "assistant"
-		}
-	],
-	"parallel_tool_calls": true,
-	"presence_penalty": 0.0,
-	"previous_response_id": null,
-	"prompt_cache_key": null,
-	"prompt_cache_retention": null,
-	"reasoning": {
-		"effort": null,
-		"summary": null
-	},
-	"safety_identifier": null,
-	"service_tier": "default",
-	"store": true,
-	"temperature": 1.0,
-	"text": {
-		"verbosity": "medium"
-	},
-	"tool_choice": "auto",
-	"tools": [],
-	"top_logprobs": 0,
-	"top_p": 1.0,
-	"truncation": "disabled",
-	"usage": {
-		"input_tokens": 100,
-		"input_tokens_details": {
-			"cached_tokens": 0
-		},
-		"output_tokens": 20,
-		"output_tokens_details": {
-			"reasoning_tokens": 0
-		},
-		"total_tokens": 120
-	},
-	"user": null,
-	"metadata": {}
-}
